fix(gen): Prevent NPE on default stream responses
The current implementation of the templated stream response `Read(byte[])` function uses the `Data io.Reader` field without first checking for `nil`.

Users who want to return a value with an empty body are forced to explicitly specify an empty `io.Reader` rather than simply ignore the field and get use from the empty struct.

This commit adds a `nil` check before using the `Data` field and, if it is `nil`, responds as if it were an empty `io.Reader` instead of panicking.

This will hopefully make this library slightly more ergonomic to use.