fix(gen): generate files with read permissions for all users
This commit updates the file permissions in the various `os.WriteFile` calls
from `600` -> `644`, as there isn't a good reason to make the files readable
only to the user who generates them.

Note that this change will not alter the generated code in any way. It only
affects the permissions on the files containing the generated code.