fix: do not fold Set-Cookie values into single header
This commit adds a special case for encoding Set-Cookie headers in responses.
Specifically, it prevents `headerParamEncoder` from folding multiple header
values into a single header value with a comma separator, as per RFC6265:

   Origin servers SHOULD NOT fold multiple Set-Cookie header fields into
   a single header field.  The usual mechanism for folding HTTP headers
   fields (i.e., as defined in [RFC2616]) might change the semantics of
   the Set-Cookie header field because the %x2C (",") character is used
   by Set-Cookie in a way that conflicts with such folding.

The comma mentioned by the RFC appears when using a cookie with an expiry
timestamp, because the timestamp format uses a comma after the day-of-week.