feat(problem+json): add support for application/problem+json
Although rfc9457 is still in proposed status, support was already
added in Java Spring environment and thus it is live in the wild.

To support this while also not yet adding a huge new mechanism just
warn that MUST declarations are currently not enforced by ogen.

To test this sprinkle problem+json into various test files.