feat: pass raw response body in UnexpectedStatusCodeError
A client sometimes encounters a truly unexpected status code from the
server for various reasons.  The client should not trash the response
away then, to instead let the caller know what was going on.  In such
cases, that bit might be useful for for instance inspection, logging,
and/or debugging purposes.