feat: handle nullable oneOf
One of the differences between OpenAPI 3.1 and 3.0 is how nullable
values are represented:

In 3.0, they are maked with `nullable: true`, and ogen handles that.

In 3.1, there are two ways:

- json schema type arrays for primitives:

```yaml
type:
  - "string"
  - "null"
```

- oneOf with a null alternative for other things (including
  primitives):

```yaml
oneOf:
  - type: "string"
  - type: "null"
```

This commits adds special hadling for latter to unlock openapi 3.1
compatible specs to be handled by ogen.

Type array support is needed for for openapi 3.1 compatability, but this
is a first step that unlocks at least some 3.1 specs to be handled by
ogen.